+++ src/controllers/console/Oauth2EncryptionController.php
@@ -4,6 +4,7 @@ namespace rhertogh\Yii2Oauth2Server\controllers\console;
 
 use rhertogh\Yii2Oauth2Server\controllers\console\base\Oauth2BaseConsoleController;
 use rhertogh\Yii2Oauth2Server\controllers\console\encryption\Oauth2EncryptionKeyUsageAction;
+use rhertogh\Yii2Oauth2Server\controllers\console\encryption\Oauth2GenerateSecretAction;
 use rhertogh\Yii2Oauth2Server\controllers\console\encryption\Oauth2RotateEncryptionKeysAction;
 use yii\helpers\ArrayHelper;
 
@@ -13,24 +14,32 @@ class Oauth2EncryptionController extends Oauth2BaseConsoleController
      * @var string|null
      */
     public $keyName = null;
+    public $secretLength = 32;
 
     /**
      * @inheritDoc
      */
     public function options($actionID)
     {
-        if (in_array($actionID, ['key-usage', 'rotate-keys'])) {
-            $options = [
+        $options = parent::options($actionID);
+        if (in_array($actionID, ['key-usage', 'rotate-keys', 'generate-secret'])) {
+            $options = ArrayHelper::merge($options, [
                 'keyName',
-            ];
+            ]);
         }
-        return ArrayHelper::merge(parent::options($actionID), $options ?? []);
+        if ($actionID === 'generate-secret') {
+            $options = ArrayHelper::merge($options, [
+                'secretLength',
+            ]);
+        }
+        return $options;
     }
 
     public function optionAliases()
     {
         return ArrayHelper::merge(parent::optionAliases(), [
             'k' => 'keyName',
+            'l' => 'secretLength',
         ]);
     }
 
@@ -42,6 +51,7 @@ class Oauth2EncryptionController extends Oauth2BaseConsoleController
         return [
             'key-usage' => Oauth2EncryptionKeyUsageAction::class,
             'rotate-keys' => Oauth2RotateEncryptionKeysAction::class,
+            'generate-secret' => Oauth2GenerateSecretAction::class,
         ];
     }
 }